+++ packages/chain-client/src/index.ts
@@ -1,4 +1,4 @@
 export { getChainAPI, destroyAll, getClient, isConnected } from "./clients.js";
 export type { Environment, ChainAPI } from "./clients.js";
-export { isInsideContainer } from "./container.js";
+export { isInsideContainer } from "@polkadot-apps/host";
 export type { ChainMeta, ConnectionMode } from "./types.js";