+++ packages/logger/src/index.ts
@@ -1,3 +1,3 @@
-export { configure } from "./configure";
-export { createLogger } from "./create-logger";
-export type { LogLevel, LogEntry, LogHandler, LoggerConfig, Logger } from "./types";
+export { configure } from "./configure.js";
+export { createLogger } from "./create-logger.js";
+export type { LogLevel, LogEntry, LogHandler, LoggerConfig, Logger } from "./types.js";