+++ packages/descriptors/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @polkadot-apps/descriptors
 
+## 0.1.2
+
+### Patch Changes
+
+- b813235: Fix npm consumer installs: remove broken `file:generated` dep from descriptors, add missing `.js` extensions to logger ESM imports.
+
+## 0.1.1
+
+### Patch Changes
+
+- 27019c9: fix: migrate npm publishing to npm_publish_automation dispatch workflow
+
 ## 0.1.0
 
 ### Minor Changes